requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
streamlit>=1.38.0
transformers>=4.44.0
torch>=2.2.0
sentence-transformers>=3.0.0
datasets>=2.20.0
scikit-learn>=1.5.0
pdfplumber>=0.11.0
pypdf>=5.0.0
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.8.0
PyYAML>=6.0.1
pytest>=8.3.0
httpx>=0.27.0
python-multipart>=0.0.9
rouge-score>=0.1.2
bert-score>=0.3.13
nltk>=3.9.0
spacy>=3.7.0
tqdm>=4.66.0
accelerate>=0.34.0
reportlab>=4.2.0
